requirements.txt
ttkbootstrap>=1.10.1
pillow>=9.0.0
python-dateutil>=2.8.2
